refactor: change esstorage to elasticsearch
